Initial node setup
The decision to use Node 12 is based on the fact that 12 is the current
LTS. Other repositories in permanent use v10, but we plan to deploy this
service separately.

The decision to use npm is that yarn represents an additional
dependency, and we don't have immediate need for things like workspaces.

Issue #1