Disable readonly-parameters lint rule
The `@typescript-eslint/prefer-readonly-parameter-types` rule does not
allow any level of mutability in function parameter definitions. This
means that neither the parameter nor any parameter attributes can be
mutable.

While this makes sense, the rule is a bit oppressive, especially when
working with express.  Two major reasons:

1. Even when wrapping a parameter with something like ReadOnly<> or
   DeepReadonly<> does not correct the issue.
2. Express expects mutable parameters, which means we cannot actually
   write valid express route handlers that are able to pass this rule.

And so, the rule is being disabled.