Test against Node.js 14
The 14.x release series of Node.js is the active long-term-support line.
We are currently using Node.js 12 in production, which is in maintenance
LTS, and is planned to be out of maintenance on 2022-04-30; we will need
to upgrade before that end-of-life date.

Start testing against Node.js 14, so that we can make sure we're ready
to upgrade once we've verified that everything is compatible.

Issue #27 Upgrade to Node 14