Upgrade Jest to resolve security vulnerability
Upgrade to the most recent version of Jest, v26.6.3[1]. This resolves a
security vulnerability in a transitive dependency, CVE-2020-7774[2] in
y18n. GitHub[3] offered to do a spot upgrade[4], but better to simply
keep on top of dependency upgrades.

There does not appear to be anything that affects us in the Jest
changelog[1], and our few tests continue to pass.

[1] https://github.com/facebook/jest/blob/master/CHANGELOG.md#2663
[2] https://nvd.nist.gov/vuln/detail/CVE-2020-7774
[3] https://github.com/advisories/GHSA-c4w7-xm78-47vh
[4] https://github.com/PermanentOrg/upload-service/pull/33