Configure logging
Set up Winston (npm[1], repo[2]), "a logger for just about everything."

Winston offers a lot of customization of both the format and the
destination of log messages. Currently, we're saving our logs to syslog
via systemd capturing stdout, so for simplicity's sake, we will continue
to do that, at least for now.

However, logging in production is very different from logging in
development. In production, systemd automatically timestamps log lines,
but when running the service via `npm start`, we do not get timestamps
by default. Additionally, while we don't need or want colors in our
production logs, they're helpful in development. Finally, making logs
easy-to-read is important in dev, but in production it's more important
to make logs easy to separate.

So: in development, show timestamps in the user's locale & time zone,
with color, and spanning multiple lines if necessary; in production,
rely on systemd to provide timestamps, do not colorize logs, and make
sure each log statement produces only a single line.

Right now, the upload-service is pretty simple, so to actually start
using the configured logging library, log when the service has
successfully started up.

[1] https://www.npmjs.com/package/winston
[2] https://github.com/winstonjs/winston

Issue #4 Bring in Winston