Improve AWS secret management
So AWS Node V2 SDK is able to pick up the credentials automatically when
set as environment variables. This means setting the shared credentials in
`~/.aws` is uneccesary and even requires another environment variable
`AWS_SDK_LOAD_CONFIG` for the SDK to use it.

This also means that, currently the image/container does not make use of the
shared credentials copy from the host.

Hence,

- We don't need the .aws shared credentials template directory (now deleted)
- We dom't need `gettext` libray which is used to dynamically update the
credentials (now excluded from image)

Signed-off-by: Fon E. Noel NFEBE <fenn25.fn@gmail.com>