Raise presigned URL timeout to 24 hours for multipart uploads
Historically, our presigned URLs for uploads have expired after 1 hour.
This makes sense for the uploads we've been doing, where the presigned
URL would be hit near instantly after it was created, but in a large
multipart upload it may be some time before each URL has been called. To
address this, this commit raises the timeout for presigned URLs involved
in a multipart upload to 24 hours.