Update eslint plugins
The eslint plugins have been awaiting updates for some time, but they've
been blocked by the unmaintained status of the airbnb typescript
ruleset. We've now decided to abandon this ruleset, so this
commit removes that dependency, updates the eslint plugins, and
reactivates dependabot for eslint. Because these changes required
significant updates to the eslintrc anyway, this commit also updates the
eslintrc to match stela's, and fixes the linting errors that resulted.