Add Prettier
This is a huge commit that changes nearly every line in the project.
This is because it introduces prettier, which is a modern tool that
pairs with eslint to handle enforcement of formatting rules.

Prettier uses tabs by default / tabs are considered the standard
baseline for code bases in TypeScript at this point.

This commit also adds prettier to our linting step and creates a new
`format` step which we can run to invoke both prettier and auto-fixable
lint rules.

Issue #511 Set up prettier