Use `eslint-config-love` ruleset
The love configuration is opinionated and robust.  Since we've disabled
airbnb this would fill the resulting gap.

We were violating some of the rules; rather than make a ton of changes
to the code base to fix all violations at once I'm just disabling the
ones that would require changes.  We can then re-enable one by one over
time.

Love includes the import plugin as part of the configuration so we no
longer need to add it directly ourselves.

Issue #578 Use the love eslint configuration preset