Enable `@typescript-eslint/no-magic-numbers`
This rule is potentially controversial but realistically it is helpful
for us to always name our numbers (unless of course we're just using "0"
to mean "first index of an array...")

Moved some constants into centralized locations and adopted an
`ALL_CAPS` naming convention for them.

This rule is much less useful for tests, where it is very common for us
to be hard coding numeric values and the context of the text makes them
not magical at all.

Issue #587 Re-enable love lint rules