Re-enable disabled rules
I think we disabled a slew of rules (many of which are good) in order to
be consistent with other projects.  I've gone through and turned many of
the rules back on.  Many of them make our code safer; in some cases
there are whitespace opinions and would suggest we default to whatever
community standard is defined by `love`.

The code doesn't violate any of the newly enabled rules at this point so
this is a simple change.

Issue #587 Re-enable love lint rules